requirements.txt
alabaster==0.7.12
altgraph==0.17
appdirs==1.4.4
astroid==2.4.2
attrs==20.1.0
autopep8==1.5.4
Babel==2.8.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==6.6
colorama==0.4.3
docutils==0.16
idna==2.10
imagesize==1.2.0
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.4
pathspec==0.8.0
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==1.4.0
Pygments==2.6.1
PyInstaller==3.6
pylint==2.5.3
pyparsing==2.4.7
pyperclip==1.8.0
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyroma==2.2
pytz==2020.1
regex==2020.7.14
requests==2.24.0
rstcheck==2.2
safety==0.5.1
six==1.15.0
snowballstemmer==2.0.0
Sphinx==1.4.9
supertools==1.0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
xmltodict==0.12.0
